requirements.txt
pyserial
pyqtgraph
PyQt6
PyQt6-tools
netCDF4
numpy
scipy
xarray